apps/personal-agi/requirements.txt
claude-agent-sdk>=0.1.19
python-dotenv>=1.0.0
structlog>=24.0.0
